refact: movendo funções de route_repo para route_stop_repo
