+++ backend/adapters/routes_repository_impl.py
@@ -1,6 +1,5 @@
 from backend.database.models.bus_departures import BusDepartures
 from backend.database.config.db_connection import DBConnection
-from backend.database.models.route_stop import RouteStop
 from backend.database.models.routes import Routes
 from backend.database.models.routes_conversion import RoutesConversion
 
@@ -8,16 +7,6 @@ class RoutesRepositoryImpl:
     
     def __init__(self):
         self.conn = DBConnection()
-
-    def return_all_routes_from_stop_impl(self, stop_id):
-        with self.conn as connection:
-            all_routes_from_stop = connection.session.query(RouteStop).filter_by(stop_id=stop_id).all()
-        return all_routes_from_stop
-      
-    def return_route_stop_by_id_impl(self, stop_id, route_id):
-        with self.conn as connection:
-            route_stop = connection.session.query(RouteStop).filter_by(stop_id=stop_id, route_id=route_id).first()
-        return route_stop
     
     def return_trips_from_route_impl(self, route_id):
         with self.conn as connection:
