+++ backend/domain/repositories/route_stop_repository.py
@@ -9,3 +9,9 @@ class RouteStopRepository:
 
     def return_all_stops_in_route(self, route_id):
         return self.repo.return_all_stops_in_route_impl(route_id)
+
+    def return_all_routes_from_stop(self, stop_id):
+        return self.repo.return_all_routes_from_stop_impl(stop_id)
+      
+    def return_route_stop_by_id(self, stop_id, route_id):
+        return self.repo.return_route_stop_by_id_impl(stop_id, route_id)
