+++ backend/domain/repositories/routes_repository.py
@@ -7,12 +7,6 @@ class RoutesRepository:
             self.repo = RoutesRepositoryImpl()
         else:
             self.repo = repo
-
-    def return_all_routes_from_stop(self, stop_id):
-        return self.repo.return_all_routes_from_stop_impl(stop_id)
-      
-    def return_route_stop_by_id(self, stop_id, route_id):
-        return self.repo.return_route_stop_by_id_impl(stop_id, route_id)
     
     def return_trips_from_route(self, route_id):
         return self.repo.return_trips_from_route_impl(route_id)
