+++ backend/domain/route/bus_schedule_impl.py
@@ -1,4 +1,5 @@
 from backend.domain.repositories.routes_repository import RoutesRepository
+from backend.domain.repositories.route_stop_repository import RouteStopRepository
 
 import datetime
 
@@ -6,14 +7,19 @@ MEAN_BUS_VELOCITY = 4.7  # velocity in m/s
 
 class BusScheduleImpl:
 
-    def __init__(self, routes_repo=None):
+    def __init__(self, routes_repo=None, route_stop_repo=None):
         if(routes_repo is None):
             self.routes_repo = RoutesRepository()
         else:
             self.routes_repo = routes_repo
 
+        if(route_stop_repo is None):
+            self.route_stop_repo = RouteStopRepository()
+        else:
+            self.route_stop_repo = route_stop_repo
+
     def get_travel_time_timedelta(self, stop_id, route_id):
-        route_stop = self.routes_repo.return_route_stop_by_id(stop_id, route_id)
+        route_stop = self.route_stop_repo.return_route_stop_by_id(stop_id, route_id)
         if not route_stop.traveled_time: 
             travel_time = route_stop.traveled_dist / MEAN_BUS_VELOCITY
         else:
